+++ packages/preview-middleware-client/test/__mock__/window.ts
@@ -1,10 +1,26 @@
 export const fetchMock = jest.fn();
 export const sapCoreMock = {
-    byId: jest.fn()
+    byId: jest.fn(),
+    getComponent: jest.fn(),
+    getConfiguration: () => ({
+        getLanguage: jest.fn()
+    })
 };
 export const sapMock = {
     ui: {
-        getCore: jest.fn().mockReturnValue(sapCoreMock)
+        getCore: jest.fn().mockReturnValue(sapCoreMock),
+        require: {
+            toUrl: jest.fn()
+        },
+        loader: {
+            config: jest.fn()
+        }
+    },
+    ushell: {
+        Container: {
+            createRenderer: jest.fn().mockReturnValue({ placeAt: jest.fn() }),
+            attachRendererCreatedEvent: jest.fn().mockImplementation((cb: () => Promise<void>) => cb())
+        }
     }
 };