+++ packages/preview-middleware/src/ui5/middleware.ts
@@ -2,9 +2,8 @@ import { LogLevel, ToolsLogger, UI5ToolingTransport } from '@sap-ux/logger';
 import type { RequestHandler } from 'express';
 import type { MiddlewareParameters, MiddlewareUtils } from '@ui5/server';
 import { FlpSandbox } from '../base/flp';
-import type { AdpPreviewConfig } from '@sap-ux/adp-tooling';
-import type { Config } from '../types';
-import { AdpPreview } from '@sap-ux/adp-tooling';
+import type { MiddlewareConfig } from '../types';
+import { AdpPreview, type AdpPreviewConfig } from '@sap-ux/adp-tooling';
 import type { ReaderCollection } from '@ui5/fs';
 
 /**
@@ -26,11 +25,17 @@ async function initAdp(
     const appVariant = await rootProject.byPath('/manifest.appdescr_variant');
     if (appVariant) {
         const adp = new AdpPreview(config, rootProject, util, logger);
-        const layer = await adp.init(JSON.parse(await appVariant.getString()));
-        flp.config.rta = {
-            layer,
-            pluginModule: 'open/ux/preview/client/adp/init'
-        };
+        const variant = JSON.parse(await appVariant.getString());
+        const layer = await adp.init(variant);
+        if (flp.rta) {
+            flp.rta.layer = layer;
+            flp.rta.options = {
+                projectId: variant.id
+            };
+            for (const editor of flp.rta.editors) {
+                editor.pluginScript ??= 'open/ux/preview/client/adp/init';
+            }
+        }
         await flp.init(adp.descriptor.manifest, adp.descriptor.name, adp.resources);
         flp.router.use(adp.descriptor.url, adp.proxy.bind(adp) as RequestHandler);
         adp.addApis(flp.router);
@@ -49,13 +54,16 @@ async function initAdp(
  * @param logger logger instance
  * @returns a router
  */
-async function createRouter({ resources, options, middlewareUtil }: MiddlewareParameters<Config>, logger: ToolsLogger) {
+async function createRouter(
+    { resources, options, middlewareUtil }: MiddlewareParameters<MiddlewareConfig>,
+    logger: ToolsLogger
+) {
     // setting defaults
     const config = options.configuration ?? {};
     config.flp ??= {};
 
     // configure the FLP sandbox based on information from the manifest
-    const flp = new FlpSandbox(config.flp, resources.rootProject, middlewareUtil, logger);
+    const flp = new FlpSandbox(config, resources.rootProject, middlewareUtil, logger);
 
     if (config.adp) {
         await initAdp(resources.rootProject, config.adp, flp, middlewareUtil, logger);
@@ -78,7 +86,7 @@ async function createRouter({ resources, options, middlewareUtil }: MiddlewarePa
  * @param params middleware configuration
  * @returns a promise for the request handler
  */
-module.exports = async (params: MiddlewareParameters<Config>): Promise<RequestHandler> => {
+module.exports = async (params: MiddlewareParameters<MiddlewareConfig>): Promise<RequestHandler> => {
     const logger = new ToolsLogger({
         transports: [new UI5ToolingTransport({ moduleName: 'preview-middleware' })],
         logLevel: params.options.configuration?.debug ? LogLevel.Debug : LogLevel.Info