+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.6.69
+
+### Patch Changes
+
+-   118692c3: update dependency http-proxy-middleware to v2.0.6
+
+## 0.6.68
+
+### Patch Changes
+
+-   Updated dependencies [676f8ba0]
+    -   @sap-ux/axios-extension@1.4.4
+
 ## 0.6.67
 
 ### Patch Changes